# api/docs/proto/openauth/v1/openauth.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "OpenAuth API",
    "description": "OpenAuth is an open-source authentication and authorization service providing user management, RBAC, and secure access control for applications.",
    "version": "v1.0.0"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/openauth/v1/auth/google/signin": {
      "post": {
        "summary": "Sign in with Google",
        "description": "Authenticate a user using a Google-issued ID token.",
        "operationId": "OpenAuth_GoogleSignIn",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SignInResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GoogleSignInRequest"
            }
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/openauth/v1/auth/logout": {
      "post": {
        "summary": "Logout user",
        "description": "Terminate user session(s).",
        "operationId": "OpenAuth_Logout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LogoutResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1LogoutRequest"
            }
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/openauth/v1/auth/refresh": {
      "post": {
        "summary": "Refresh access token",
        "description": "Generate new access token using refresh token.",
        "operationId": "OpenAuth_RefreshToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RefreshTokenResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1RefreshTokenRequest"
            }
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/openauth/v1/auth/signin": {
      "post": {
        "summary": "Sign in user",
        "description": "Authenticate a user and create a new session.",
        "operationId": "OpenAuth_SignIn",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SignInResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SignInRequest"
            }
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/openauth/v1/auth/signin-with-token": {
      "post": {
        "summary": "Sign in with login token",
        "description": "Authenticate using a single-use login token. Returns the existing session without creating a new one.",
        "operationId": "OpenAuth_SignInWithLoginToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SignInResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SignInWithLoginTokenRequest"
            }
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/openauth/v1/auth/validate": {
      "post": {
        "summary": "Validate access token",
        "description": "Check if an access token is valid and active.",
        "operationId": "OpenAuth_ValidateToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ValidateTokenResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ValidateTokenRequest"
            }
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/openauth/v1/users/signup": {
      "post": {
        "summary": "Sign up a new user",
        "description": "Create a new user account in the system.",
        "operationId": "OpenAuth_SignUp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SignUpResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "SignUpRequest for user registration - only authentication credentials.\nAt least one of username, email, or phone must be provided; if username\nis omitted, one is generated automatically from the email or phone.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SignUpRequest"
            }
          }
        ],
        "tags": [
          "User Registration"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1GoogleSignInRequest": {
      "type": "object",
      "properties": {
        "idToken": {
          "type": "string"
        },
        "metadata": {
          "$ref": "#/definitions/v1SignInMetadata"
        },
        "profiles": {
          "type": "boolean"
        },
        "includePermissions": {
          "type": "boolean"
        }
      },
      "title": "GoogleSignInRequest authenticates a user via a Google-issued ID token"
    },
...
    "v1LogoutRequest": {
      "type": "object",
      "properties": {
        "sessionId": {
          "type": "string",
          "title": "Specific session to logout"
        },
        "allSessions": {
          "type": "boolean",
          "title": "Logout from all devices"
        }
      },
      "title": "LogoutRequest to end user session"
    },
    "v1LogoutResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "sessionsTerminated": {
          "type": "integer",
          "format": "int32",
          "title": "Number of sessions ended"
        }
      },
      "title": "LogoutResponse"
    },
...
    "v1RefreshTokenRequest": {
      "type": "object",
      "properties": {
        "refreshToken": {
          "type": "string"
        },
        "deviceId": {
          "type": "string"
        },
        "profiles": {
          "type": "boolean"
        },
        "includePermissions": {
          "type": "boolean"
        }
      },
      "title": "RefreshTokenRequest to refresh access token"
    },
    "v1RefreshTokenResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string",
          "title": "New refresh token (rotation)"
        },
        "expiresAt": {
          "type": "string",
          "format": "int64"
        },
        "refreshExpiresAt": {
          "type": "string",
          "format": "int64"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "RefreshTokenResponse with new tokens"
    },
...
    "v1SignInMetadata": {
      "type": "object",
      "properties": {
        "deviceId": {
          "type": "string",
          "title": "Unique device identifier"
        },
        "deviceName": {
          "type": "string",
          "title": "Human-readable device name"
        },
        "deviceType": {
          "type": "string",
          "title": "web, mobile, desktop, tablet"
        },
        "lat": {
          "type": "number",
          "format": "double"
        },
        "long": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "v1SignInRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "otp": {
          "type": "string"
        },
        "rememberMe": {
          "type": "boolean"
        },
        "metadata": {
          "$ref": "#/definitions/v1SignInMetadata"
        },
        "profiles": {
          "type": "boolean"
        },
        "includePermissions": {
          "type": "boolean"
        },
        "verificationId": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "SignInRequest for user authentication"
    },
    "v1SignInResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "expiresAt": {
          "type": "string",
          "format": "int64",
          "title": "Access token expiration (Unix timestamp)"
        },
        "refreshExpiresAt": {
          "type": "string",
          "format": "int64",
          "title": "Refresh token expiration (Unix timestamp)"
        },
        "user": {
          "$ref": "#/definitions/v1User"
        },
        "sessionId": {
          "type": "string",
          "title": "Session UUID for tracking"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "SignInResponse with authentication tokens and user data"
    },
    "v1SignInWithLoginTokenRequest": {
      "type": "object",
      "properties": {
        "loginToken": {
          "type": "string"
        },
        "metadata": {
          "$ref": "#/definitions/v1SignInMetadata"
        },
        "profiles": {
          "type": "boolean"
        },
        "includePermissions": {
          "type": "boolean"
        }
      },
      "title": "SignInWithLoginTokenRequest - sign in using a single-use login token"
    },
    "v1SignUpRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "description": "SignUpRequest for user registration - only authentication credentials.\nAt least one of username, email, or phone must be provided; if username\nis omitted, one is generated automatically from the email or phone."
    },
    "v1SignUpResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/v1User"
        },
        "message": {
          "type": "string"
        },
        "emailVerificationRequired": {
          "type": "boolean"
        },
        "phoneVerificationRequired": {
          "type": "boolean"
        }
      },
      "title": "SignUpResponse after successful registration - only user data"
    },
...
    "v1User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "uuid": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "name": {
          "type": "string",
          "title": "Display name for the user"
        },
        "avatarUrl": {
          "type": "string",
          "title": "URL to user's avatar image"
        },
        "emailVerified": {
          "type": "boolean"
        },
        "phoneVerified": {
          "type": "boolean"
        },
        "deactivated": {
          "type": "boolean"
        },
        "isActive": {
          "type": "boolean"
        },
        "isLocked": {
          "type": "boolean"
        },
        "failedLoginAttempts": {
          "type": "integer",
          "format": "int32"
        },
        "lastLoginAt": {
          "type": "string",
          "format": "int64"
        },
        "passwordChangedAt": {
          "type": "string",
          "format": "int64"
        },
        "createdAt": {
          "type": "string",
          "format": "int64"
        },
        "updatedAt": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "User represents a user account in the system"
    },
...
    "v1ValidateTokenRequest": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        }
      },
      "title": "ValidateTokenRequest to check token validity"
    },
    "v1ValidateTokenResponse": {
      "type": "object",
      "properties": {
        "valid": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/v1User",
          "title": "User info if token is valid"
        },
        "expiresAt": {
          "type": "string",
          "format": "int64",
          "title": "Token expiration"
        }
      },
      "title": "ValidateTokenResponse"
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "description": "Bearer token for authentication. Format: 'Bearer \u003ctoken\u003e'",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}